Refactored Ingredient Hierarchy
In short, I've stripped out the type hierarchy that defined ingredients. As cool as it was, it was creating more problems than it solved. Now everything is an ingredient, with a superType, Type, and subType attribute. These may prove unnecessary later, as it has become clear that it's really all about the lists.